Move some actions from actions/send.ts to actions/transaction.ts
